scraper/requirements.txt
aiohttp
aiodns
async-timeout
feedparser
beautifulsoup4
psycopg2-binary
python-dateutil